Fix some errors

+++ src/unattend_my_iso/core/files/file_manager.py
@@ -1,6 +1,7 @@
 import os
+from os.path import isdir, isfile
 import shutil
-
+import subprocess
 from unattend_my_iso.helpers.logging import log_error
 
 
@@ -9,21 +10,49 @@ class UmiFileManager:
     def __init__(self):
         pass
 
-    def copy_file(self, src: str, dst: str) -> bool:
+    def rm(self, src: str) -> bool:
         try:
-            shutil.copy(src, dst)
+            if os.path.exists(src):
+                if isfile(src):
+                    os.remove(src)
+                elif isdir(src):
+                    shutil.rmtree(src)
+                else:
+                    log_error("Cant delete unknown file object")
         except Exception as exe:
             log_error(f"Error on copy_file: {exe}")
             return False
         return True
 
-    def copy_folder(self, src: str, dst: str) -> bool:
+    def mv(self, src: str, dst: str) -> bool:
         try:
-            if os.path.exists(dst):
-                shutil.rmtree(dst)
-            shutil.copytree(src, dst, dirs_exist_ok=True, symlinks=True)
+            if os.path.exists(src):
+                shutil.move(src, dst)
         except Exception as exe:
-            log_error(f"Error on copy_folder: {exe}")
+            log_error(f"Error on copy_file: {exe}")
+            return False
+        return True
+
+    def cp(self, src: str, dst: str) -> bool:
+        try:
+            if os.path.exists(src):
+                if isfile(src):
+                    shutil.copy(src, dst)
+                elif isdir(src):
+                    shutil.copytree(src, dst, dirs_exist_ok=True, symlinks=True)
+                else:
+                    log_error("Cant delete unknown file object")
+        except Exception as exe:
+            log_error(f"Error on copy_file: {exe}")
+            return False
+        return True
+
+    def copy_folder_iso(self, src: str, dst: str) -> bool:
+        try:
+            self.rm(dst)
+            subprocess.run(["cp", "-r", src, dst])
+        except Exception as exe:
+            log_error(f"Error on copy_folder_iso: {exe}")
             return False
         return True
